trac_migration/requirements.txt
# Common
python-dotenv

# For issue migration
PyGithub

# For the wiki
beautifulsoup4
bleach
lxml
